+++ src/preprocess/Preprocess.java
@@ -21,19 +21,19 @@ public class Preprocess {
 	}
 	
 	public static void  vertical_trans(Image image) {
-		byte[][] temp = new byte[image.numcomps][image.height*image.width];
+		float[][] temp = new float[image.numcomps][image.height*image.width];
 		int index = 0;
 		
 		for (index = 0; index < image.width*image.height; index++) {
-			temp[0][index] = (byte)(0.25*image.comps[0][index] + 0.5*image.comps[1][index] + 0.25*image.comps[2][index]);
+			temp[0][index] = (float)(0.25*image.comps[0][index] + 0.5*image.comps[1][index] + 0.25*image.comps[2][index]);
 		}
 		
 		for (index = 0; index < image.width*image.height; index++) {
-			temp[1][index] = (byte)(image.comps[2][index] - image.comps[1][index]);
+			temp[1][index] = (float)(image.comps[2][index] - image.comps[1][index]);
 		}
 		
 		for (index = 0; index < image.width*image.height; index++) {
-			temp[2][index] = (byte)(image.comps[0][index] - image.comps[1][index]);
+			temp[2][index] = (float)(image.comps[0][index] - image.comps[1][index]);
 		}
 		
 		image.comps = temp;
